+++ src/xmap.cpp
@@ -8,31 +8,32 @@ namespace edm
 {
 
 void xmap(CrossMap &result, const Dataset &ds, const TimeSeries &library,
-          const std::vector<int> &edims, int E_max, int tau, int Tp)
+          const std::vector<uint32_t> &edims, uint32_t E_max, int32_t tau,
+          int32_t Tp)
 {
     std::vector<LUT> luts;
 
     // Allocate kNN tables
-    for (int E = 1; E <= E_max; E++) {
+    for (uint32_t E = 1; E <= E_max; E++) {
         luts.push_back(LUT(ds.extent(0) - (E - 1) * tau, E + 1));
     }
 
     LUT tmp_lut(ds.extent(0), ds.extent(0));
 
     // Compute kNN tables for all E
-    for (int E = 1; E <= E_max; E++) {
+    for (uint32_t E = 1; E <= E_max; E++) {
         knn(library, library, luts[E - 1], tmp_lut, E, tau, Tp, E + 1);
         normalize_lut(luts[E - 1]);
     }
 
     // Group time series by their optimal embedding dimensions
     std::vector<std::vector<uint32_t>> groups(E_max);
-    for (size_t i = 0; i < ds.extent(1); i++) {
+    for (uint32_t i = 0; i < ds.extent(1); i++) {
         groups[edims[i] - 1].push_back(i);
     }
 
     // Perform lookups
-    for (int E = 1; E <= E_max; E++) {
+    for (uint32_t E = 1; E <= E_max; E++) {
         if (!groups[E - 1].size()) {
             continue;
         }
@@ -52,13 +53,13 @@ void xmap(CrossMap &result, const Dataset &ds, const TimeSeries &library,
             Kokkos::TeamPolicy<>(targets.size(), Kokkos::AUTO)
                 .set_scratch_size(0, Kokkos::PerTeam(scratch_size)),
             KOKKOS_LAMBDA(const Kokkos::TeamPolicy<>::member_type &member) {
-                int tj = targets(member.league_rank());
+                uint32_t tj = targets(member.league_rank());
 
                 ScratchTimeSeries scratch(member.team_scratch(0), ds.extent(0));
 
                 Kokkos::parallel_for(
                     Kokkos::TeamThreadRange(member, ds.extent(0)),
-                    [=](int i) { scratch(i) = ds(i, tj); });
+                    [=](uint32_t i) { scratch(i) = ds(i, tj); });
 
                 member.team_barrier();
 
@@ -66,12 +67,12 @@ void xmap(CrossMap &result, const Dataset &ds, const TimeSeries &library,
 
                 Kokkos::parallel_reduce(
                     Kokkos::TeamThreadRange(member, distances.extent(0)),
-                    [=](int i, CorrcoefState &upd) {
+                    [=](uint32_t i, CorrcoefState &upd) {
                         float pred = 0.0f;
 
                         Kokkos::parallel_reduce(
                             Kokkos::ThreadVectorRange(member, E + 1),
-                            [=](int &e, float &p) {
+                            [=](uint32_t &e, float &p) {
                                 p += scratch(indices(i, e)) * distances(i, e);
                             },
                             pred);
